// client-next/src/api/models/PaginatedUser.ts
export interface PaginatedUser {
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    currentPage?: number;
    /**
     * 
     * @type {string}
     * @memberof PaginatedUser
     */
    firstPageUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    from?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    lastPage?: number;
    /**
     * 
     * @type {string}
     * @memberof PaginatedUser
     */
    lastPageUrl?: string;
    /**
     * 
     * @type {Array<PaginationLinksInner>}
     * @memberof PaginatedUser
     */
    links?: Array<PaginationLinksInner>;
    /**
     * 
     * @type {string}
     * @memberof PaginatedUser
     */
    nextPageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PaginatedUser
     */
    path?: string;
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    perPage?: number;
    /**
     * 
     * @type {string}
     * @memberof PaginatedUser
     */
    prevPageUrl?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    to?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginatedUser
     */
    total?: number;
    /**
     * 
     * @type {Array<User1Inner>}
     * @memberof PaginatedUser
     */
    data?: Array<User1Inner>;
}