// client-next/src/api/models/StoreUserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Practice API Documentation
 * Practice API description
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: admin@admin.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StoreUserRequest
 */
export interface StoreUserRequest {
    /**
     * Name of the user
     * @type {string}
     * @memberof StoreUserRequest
     */
    name: string;
    /**
     * Email of the user
     * @type {string}
     * @memberof StoreUserRequest
     */
    email: string;
    /**
     * Password of the user
     * @type {string}
     * @memberof StoreUserRequest
     */
    password: string;
}

/**
 * Check if a given object implements the StoreUserRequest interface.
 */
export function instanceOfStoreUserRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;

    return isInstance;
}

export function StoreUserRequestFromJSON(json: any): StoreUserRequest {
    return StoreUserRequestFromJSONTyped(json, false);
}

export function StoreUserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoreUserRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'email': json['email'],
        'password': json['password'],
    };
}

export function StoreUserRequestToJSON(value?: StoreUserRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'email': value.email,
        'password': value.password,
    };
}