// client-next/src/api/models/User.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Practice API Documentation
 * Practice API description
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: admin@admin.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * User model
 * @export
 * @interface User
 */
export interface User {
    /**
     * ID of the user
     * @type {any}
     * @memberof User
     */
    id?: any | null;
    /**
     * Name of the user
     * @type {any}
     * @memberof User
     */
    name?: any | null;
    /**
     * Email of the user
     * @type {any}
     * @memberof User
     */
    email?: any | null;
    /**
     * Email verification
     * @type {any}
     * @memberof User
     */
    emailVerifiedAt?: any | null;
    /**
     * Password of the user
     * @type {any}
     * @memberof User
     */
    password?: any | null;
}

/**
 * Check if a given object implements the User interface.
 */
export function instanceOfUser(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function UserFromJSON(json: any): User {
    return UserFromJSONTyped(json, false);
}

export function UserFromJSONTyped(json: any, ignoreDiscriminator: boolean): User {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'emailVerifiedAt': !exists(json, 'email_verified_at') ? undefined : json['email_verified_at'],
        'password': !exists(json, 'password') ? undefined : json['password'],
    };
}

export function UserToJSON(value?: User | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'email': value.email,
        'email_verified_at': value.emailVerifiedAt,
        'password': value.password,
    };
}